rename Arcada subtype to Adafruit_SPITFT, make a header file for each board type
